Clean up duplicated gameplay helpers
